Acts 1 2 i 3 optimitzades

+++ CODE INSITUT CFGS/Programacio/Java/Ud3Act1MarcMas/src/main/java/org/yourcompany/yourproject/Ud3Act1MarcMas.java
@@ -14,7 +14,7 @@ public class Ud3Act1MarcMas {
     public static void main(String[] args) {
         Scanner s = new Scanner(System.in);
         System.out.println("Introdueixi un numero"); //deman el primer nombre per consola i l'agaf com a double
-        Double n = s.nextDouble();
+        float n = s.nextFloat();
         while (n != 0) { //si n NO es 0, entram
             if (n % 2 == 0) { //si el nombre es divisible entre 0 es parell
                 System.out.println("El nombre es Parell");
@@ -23,7 +23,7 @@ public class Ud3Act1MarcMas {
             }
             System.out.println("El teu nombre al quadrat es " + Math.pow(n, 2)); //calcul el nombre al quadrat
             System.out.println("Introdueixi un numero"); //deman el seguent nombre 
-            n = s.nextDouble();
+            n = s.nextFloat();
         }
     }
 }
