+++ CODE INSITUT CFGS/Programacio/Java/Ud3Act3MarcMas/src/main/java/org/yourcompany/yourproject/Ud3Act3MarcMas.java
@@ -14,14 +14,14 @@ public class Ud3Act3MarcMas {
     public static void main(String[] args) {
         Scanner s = new Scanner(System.in);
         System.out.println("Introdueixi un nombre"); //deman el primer nombre per consola i l'agaf com a int
-        Integer n = s.nextInt();
-        Integer rondes = 0; //inicialitzam variables
+        Byte n = s.nextByte();
+        Byte rondes = 0; //inicialitzam variables
         Integer punts = 0;
         Integer solucio = (int) (Math.random() * 100); //Math.random genera un nombre entre 0.0 i 1.0, el multiplicare per 100 per tenir un valor entre 0 i 100
         if (solucio == 0) {
             solucio = 1; //per evitar que el random doni 0
         }
-        while (n != solucio) { //si n no es igual que la solucio, entram
+        while ((int) n != solucio) { //si n no es igual que la solucio, entram
             if (n > solucio) {
                 System.out.println("El teu nombre es mes alt que la solucio");
                 punts += (n-solucio);
@@ -31,7 +31,7 @@ public class Ud3Act3MarcMas {
             }
             rondes++; //sumam una ronda
             System.out.println("Introdueixi un altre"); //deman el primer nombre per consola i l'agaf com a int
-            n = s.nextInt();
+            n = s.nextByte();
         }
         //un pic encertam el nombre, surtim
         System.out.println("Enhorabona! La teva puntuacio es de "+rondes+" Rondes - "+punts+" Punts");
